Simplificados fillReportDetails con streams y añadido clear al grafo Revisar aún así el de Road porque para acceder a la lista de vehículos se accede al InnerList del MultiTreeMap, que parece que tiene un get con O(n), con lo cual probablemente el algoritmo sea cuadrático. Vendría bien poder copiar la lista quizá, que también es O(n), pero sólo una vez al principio y ya habría un get O(1).

+++ src/main/java/es/ucm/fdi/model/LaneRoad.java
@@ -16,8 +16,7 @@ public class LaneRoad extends Road {
 
   @Override
   protected int calculateBaseSpeed() {
-    return (int) Math.min(maxSpeed,
-        maxSpeed * lanes / Math.max(vehicleList.sizeOfValues(), 1) + 1);
+    return Math.min(maxSpeed, maxSpeed * lanes / Math.max(vehicleList.sizeOfValues(), 1) + 1);
   }
 
   @Override
